pyproject.toml
[tool.poetry]
name = "lucius"
version = "0.1.0"
description = "Study assistant powered by open ML models."
authors = ["malyvsen <5940672+malyvsen@users.noreply.github.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
faster-whisper = "^0.3.0"
tqdm = "^4.65.0"
click = "^8.1.3"
transformers = "^4.27.3"
torch = "^2.0.0"
pypdf = "^3.7.0"
pillow = "^9.4.0"
numpy = "^1.24.2"
open-clip-torch = "^2.16.0"
torchvision = "^0.15.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
jupyter = "^1.0.0"
plotly = "^5.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
